:bug: 檢查玩家是否正在玩遊戲，修改賞金資料型態為Int

+++ utopia-gamification/src/main/kotlin/tw/waterballsa/utopia/utopiagamification/repositories/mongodb/repositoryimpl/MongodbPlayerRepository.kt
@@ -36,7 +36,7 @@ class MongodbPlayerRepository(
         name,
         exp.toInt(),
         level.toInt(),
-        bounty.toInt(),
+        bounty,
         joinDate,
         latestActivateDate,
         levelUpgradeDate,
